pyproject.toml
[project]
name = "viralbite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "google-api-python-client>=2.193.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "matplotlib>=3.9.0",
    "langchain-google-vertexai>=1.0.5",
    "pandas>=2.2.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.2.2",
    "youtube-transcript-api>=0.6.2",
    "uvicorn>=0.30.0",
    "vaderSentiment>=3.3.2",
]
